refactor: Move ingester sort key to schema sort key
This logic isn't actually ingester specific
