+++ ingester/src/compact.rs
@@ -1,9 +1,6 @@
 //! This module is responsible for compacting Ingester's data
 
-use crate::{
-    data::{PersistingBatch, QueryableBatch},
-    sort_key::{adjust_sort_key_columns, compute_sort_key},
-};
+use crate::data::{PersistingBatch, QueryableBatch};
 use arrow::record_batch::RecordBatch;
 use data_types2::{NamespaceId, PartitionInfo};
 use datafusion::{error::DataFusionError, physical_plan::SendableRecordBatchStream};
@@ -15,7 +12,7 @@ use query::{
     util::compute_timenanosecond_min_max,
     QueryChunk, QueryChunkMeta,
 };
-use schema::sort::SortKey;
+use schema::sort::{adjust_sort_key_columns, compute_sort_key, SortKey};
 use snafu::{ResultExt, Snafu};
 use std::sync::Arc;
 use time::{Time, TimeProvider};
