+++ ingester/src/lib.rs
@@ -24,7 +24,6 @@ mod poison;
 pub mod querier_handler;
 pub mod query;
 pub mod server;
-pub mod sort_key;
 pub mod stream_handler;
 
 #[cfg(test)]
